RUMM-674 Shopist randomizes IP addresses
